Refactor controller function to verify user can update or delete a variable and remove try-catch blocks because promise based router already catches any unhandled error

+++ src/controllers/variableController.js
@@ -1,53 +1,55 @@
-const { Variable, Formula } = require('../models');
-const { validate } = require('../services/authService');
+const { Variable } = require('../models');
 const { getIdIndicadorRelatedTo } = require('../services/indicadorService');
+const { isUsuarioAssignedToIndicador } = require('../services/usuarioIndicadorService');
 
 const updateVariable = async (req, res, next) => {
   const { idVariable, ...values } = req.matchedData;
   const rol = req.rol;
-  const idUsuario = req.sub;
-  try {
-    const variable = await Variable.findOne({
-      where: { id: idVariable },
-      attributes: ['idFormula'],
-      raw: true
-    })
-    const idIndicador = await getIdIndicadorRelatedTo(Formula, variable.idFormula);
-    return validate(
-      { rol, idUsuario, idIndicador },
-      async () => {
-        const updatedRows = await Variable.update(values, { where: { id: idVariable } });
-        return updatedRows > 0 ? res.sendStatus(204) : res.sendStatus(400);
-      },
-      () => res.status(403).send('No tienes permiso para actualizar esta variable')
-    )
-  } catch (err) {
-    next(err);
+  if (rol === 'ADMIN') {
+    await Variable.update(values, { where: { id: idVariable } });
+    return res.sendStatus(204);
   }
+
+  const variable = await Variable.findOne({
+    where: { id: idVariable },
+    attributes: ['idFormula'],
+    raw: true
+  })
+
+  const idIndicador = await getIdIndicadorRelatedTo('Formula', variable.idFormula);
+  const userIsAssignedToIndicador = await isUsuarioAssignedToIndicador(req.sub, idIndicador)
+  if (!userIsAssignedToIndicador) {
+    return res.status(403).send('No tienes permiso para realizar esta operación')
+  }
+
+  await Variable.update(values, { where: { id: idVariable } });
+  return res.sendStatus(204);
+
 }
 
 const deleteVariable = async (req, res, next) => {
   const { idVariable } = req.matchedData;
   const rol = req.rol;
-  const idUsuario = req.sub;
-  try {
-    const variable = await Variable.findOne({
-      where: { id: idVariable },
-      attributes: ['idFormula'],
-      raw: true
-    })
-    const idIndicador = await getIdIndicadorRelatedTo(Formula, variable.idFormula);
-    return validate(
-      { rol, idUsuario, idIndicador },
-      async () => {
-        const destroyed = await Variable.destroy({ where: { id: idVariable } });
-        return destroyed > 0 ? res.sendStatus(204) : res.sendStatus(400);
-      },
-      () => res.status(403).send('No tienes permiso para eliminar esta variable')
-    )
-  } catch (err) {
-    next(err)
+
+  if (rol === 'ADMIN') {
+    await Variable.destroy({ where: { id: idVariable } });
+    return res.sendStatus(204);
   }
+
+  const variable = await Variable.findOne({
+    where: { id: idVariable },
+    attributes: ['idFormula'],
+    raw: true
+  })
+
+  const idIndicador = await getIdIndicadorRelatedTo('Formula', variable.idFormula);
+  const userIsAssignedToIndicador = await isUsuarioAssignedToIndicador(req.sub, idIndicador)
+  if (!userIsAssignedToIndicador) {
+    return res.status(403).send('No tienes permiso para realizar esta operación')
+  }
+
+  await Variable.destroy({ where: { id: idVariable } });
+  return res.sendStatus(204);
 }
 
 module.exports = {
